requirements.txt
aiogram>=3.10.0
aiohttp>=3.10.0
PyGithub>=2.0.0
python-dotenv>=1.0.0
